Add terraforming-control-plane
[finishes #162563560](https://www.pivotaltracker.com/story/show/162563560)

Configure and Allow Connection to External DB

Based on how RelEng does this, terraform calls a small bash script
in terraforming-control-plane. The script creates an SSH tunnel to the
ops man vm and runs psql commands. It also cleans up after itself.

The multiple `terraform apply` executions with and without the override
is no longer necessary and the DB is 'always' only accessable from the
internal network.

We manually tested up to apply-changes with the control plane tile and
everything seemed to work.

Database script now used to provision the separate databases required
to run concourse (atc,uaa,etc).

Co-authored-by: Chris Dutra <cdutra@pivotal.io>
Co-authored-by: Christopher Hunter <hunter.christopher@icloud.com>
Co-authored-by: Sara Montecino <saramontecino@outlook.com>
Co-authored-by: Nick Mahoney <nmahoney@pivotal.io>
Co-authored-by: Zach Gershman <zgershman@pivotal.io>