Make use of ALBs optional in control plane.
US Gov Cloud does not allow for ALBs. This PR allows for that, by
setting a new flag `use_alb` that, when set to false, will create
2 new NLBs for credhub and uaa, rather than a single ALB, with a
similar user experience. The default has been set to true so that
existing tfvars files will continue to work as it has before.